Alumni Blog Created

+++ student-user/includes/content.php
@@ -83,9 +83,11 @@
 
   <main id="main">
 
+
     <!--==========================
       Featured Services Section
     ============================-->
+    <!--
     <section id="featured-services">
       <div class="container">
         <div class="row">
@@ -110,7 +112,9 @@
 
         </div>
       </div>
-    </section><!-- #featured-services -->
+    </section> -->
+<!-- #featured-services -->
+
 
     <!--==========================
       About Us Section
@@ -127,9 +131,9 @@
 
           <div class="col-md-4 wow fadeInUp">
             <div class="about-col">
-              <div class="img">
+              <div class="card">
                 <img src="img/about-mission.jpg" alt="" class="img-fluid">
-                <div class="icon"><i class="ion-ios-speedometer-outline"></i></div>
+                <div class="icon"><a href="#"><i class="ion-ios-speedometer-outline"></i></a></div>
               </div>
               <h2 class="title"><a href="#">Our Mission</a></h2>
               <p>
@@ -176,10 +180,10 @@
       <div class="container">
 
         <header class="section-header wow fadeInUp">
-          <h3>Services</h3>
+          <h3>Career Opportunity</h3>
           <p>Laudem latine persequeris id sed, ex fabulas delectus quo. No vel partiendo abhorreant vituperatoribus, ad pro quaestio laboramus. Ei ubique vivendum pro. At ius nisl accusam lorenta zanos paradigno tridexa panatarel.</p>
         </header>
-
+<!--
         <div class="row">
 
           <div class="col-lg-4 col-md-6 box wow bounceInUp" data-wow-duration="1.4s">
@@ -214,24 +218,28 @@
           </div>
 
         </div>
-
+-->
       </div>
     </section><!-- #services -->
 
     <!--==========================
       Call To Action Section
     ============================-->
+    <!--
     <section id="call-to-action" class="wow fadeIn">
       <div class="container text-center">
         <h3>Call To Action</h3>
         <p> Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
         <a class="cta-btn" href="#">Call To Action</a>
       </div>
-    </section><!-- #call-to-action -->
+    </section>
+    -->
+    <!-- #call-to-action -->
 
     <!--==========================
       Skills Section
     ============================-->
+    <!--
     <section id="skills">
       <div class="container">
 
@@ -270,10 +278,11 @@
 
       </div>
     </section>
-
+-->
     <!--==========================
       Facts Section
     ============================-->
+    <!--
     <section id="facts"  class="wow fadeIn">
       <div class="container">
 
@@ -311,9 +320,10 @@
         </div>
 
       </div>
-    </section><!-- #facts -->
+    </section>
+    -->
+    <!-- #facts -->
 
- 
  <!--==========================
       Portfolio Section
     ============================-->
@@ -347,7 +357,7 @@
 //alumni blog
 $query = "SELECT uposts.*, categories.name as category_name, users.name as user_name FROM `uposts` 
             LEFT JOIN categories ON uposts.category_id=categories.id 
-            LEFT JOIN users ON uposts.user_id=users.id";
+            LEFT JOIN users ON uposts.user_id=users.id ";
     $posts = $db->getData($query);
     $user_id= $_SESSION['id'];
 
@@ -355,14 +365,14 @@ $query = "SELECT uposts.*, categories.name as category_name, users.name as user_
   //student blog
     $query1 = "SELECT sposts.*, categories.name as category_name, students.name as student_name FROM `sposts` 
             LEFT JOIN categories ON sposts.category_id=categories.id 
-            LEFT JOIN students ON sposts.student_id=students.id";
+            LEFT JOIN students ON sposts.student_id=students.id ORDER BY id DESC";
     $posts1 = $db->getData($query1);
     $student_id= $_SESSION['id'];
 
     //admin blog
     $query2 = "SELECT posts.*, categories.name as category_name, admins.name as admin_name FROM `posts` 
             LEFT JOIN categories ON posts.category_id=categories.id 
-            LEFT JOIN admins ON posts.admin_id=admins.id";
+            LEFT JOIN admins ON posts.admin_id=admins.id ";
     $posts2 = $db->getData($query2);
 ?>
         <div class="row ">
@@ -380,7 +390,7 @@ $query = "SELECT uposts.*, categories.name as category_name, users.name as user_
                             <?php echo $post['category_name'];  
                                     if($post['user_id']==$user_id){?>  
                                     <a href="edit-post.php?edit=<?php echo $post['id']; ?>" style="float:right;" class="btn btn-success btn-sm"> <img src="../alumni-user/img/edit.png" alt="Avatar" >Edit</a>
-                                    <a href="delete-post.php?delete=<?php echo $post['id']; ?>" style="float:right;" class="btn btn-danger btn-sm"><img src="../alumni-user/img/delete.png" alt="Avatar" >Delete</a>
+                                    <a onclick="return confirm('Do You Want to delete this Blog?')" href="delete-post.php?delete=<?php echo $post['id']; ?>" style="float:right;" class="btn btn-danger btn-sm"><img src="../alumni-user/img/delete.png" alt="Avatar" >Delete</a>
                                         
                                 <?php }?>
                         </div>
@@ -420,7 +430,7 @@ $query = "SELECT uposts.*, categories.name as category_name, users.name as user_
                            
                                     if($post1['student_id']==$student_id){?>  
                                   <a href="edit-post.php?edit=<?php echo $post1['id']; ?>" style="float:right;" class="btn btn-success btn-sm"><i class="fas fa-user-edit"></i>Edit</a>
-                                    <a href="delete-post.php?delete=<?php echo $post1['id']; ?>" style="float:right;" class="btn btn-danger btn-sm"><i class="fas fa-trash-alt"></i>Delete</a>
+                                    <a onclick="return confirm('Do You Want to delete this Blog?')" href="delete-post.php?delete=<?php echo $post1['id']; ?>" style="float:right;" class="btn btn-danger btn-sm"><i class="fas fa-trash-alt"></i>Delete</a>
                                         
                                 <?php }?>
                         </div>
@@ -635,9 +645,11 @@ $query = "SELECT uposts.*, categories.name as category_name, users.name as user_
 
 
     </section><!-- #portfolio -->
+
     <!--==========================
       Clients Section
     ============================-->
+    <!--
     <section id="clients" class="wow fadeInUp">
       <div class="container">
 
@@ -657,11 +669,13 @@ $query = "SELECT uposts.*, categories.name as category_name, users.name as user_
         </div>
 
       </div>
-    </section><!-- #clients -->
+    </section>
+    --><!-- #clients -->
 
     <!--==========================
       Clients Section
     ============================-->
+    <!--
     <section id="testimonials" class="section-bg wow fadeInUp">
       <div class="container">
 
@@ -729,7 +743,9 @@ $query = "SELECT uposts.*, categories.name as category_name, users.name as user_
         </div>
 
       </div>
-    </section><!-- #testimonials -->
+    </section>
+    -->
+    <!-- #testimonials -->
 
     <!--==========================
       Team Section
@@ -737,7 +753,7 @@ $query = "SELECT uposts.*, categories.name as category_name, users.name as user_
     <section id="team">
       <div class="container">
         <div class="section-header wow fadeInUp">
-          <h3>Team</h3>
+          <h3>Events</h3>
           <p>Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque</p>
         </div>
 
@@ -823,6 +839,7 @@ $query = "SELECT uposts.*, categories.name as category_name, users.name as user_
     <!--==========================
       Contact Section
     ============================-->
+    <!--
     <section id="contact" class="section-bg wow fadeInUp">
       <div class="container">
 
@@ -886,3 +903,4 @@ $query = "SELECT uposts.*, categories.name as category_name, users.name as user_
         </div>
 
       </div>
+      -->
